Gaffer : Added TaskContextProcessor.
This simplifies the implementation of the Wedge node, and paves the way for other similar nodes.
